fix: errors

+++ segment/src/app/cart/cart.component.ts
@@ -2,12 +2,14 @@ import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../service/api.service';
 import { CartItem } from '../models/cart';
 import { AuthService } from '../service/auth.service';
+import jwt_decode from 'jwt-decode';
 
 @Component({
   selector: 'app-cart',
   templateUrl: './cart.component.html',
   styleUrls: ['./cart.component.css']
 })
+
 export class CartComponent implements OnInit {
   cartItems: CartItem[] = [];
 
@@ -17,6 +19,7 @@ export class CartComponent implements OnInit {
 
   ngOnInit() {
     this.currentUser = this.authService.getCurrentUser()
+    
     // this.getCartItems();
   }
 
@@ -28,11 +31,21 @@ export class CartComponent implements OnInit {
   }
 
   removeFromCart(item: CartItem): void {
+  const token : string | null = localStorage.getItem('token');
+  if (token !== null) {
+    // делать что-то с токеном
+  
+  const decodedToken: any  = jwt_decode(token);
+  
+  const userId = decodedToken.userId;
+  
+
     // Call the API service to remove the item from the user's cart
     this.apiService.removeFromCart(userId,item.id).subscribe(() => {
       // Remove the item from the cartItems array
       const index = this.cartItems.indexOf(item);
       this.cartItems.splice(index, 1);
     });
+  }
   }
 }
